+++ tests/testthat/test-plot.R
@@ -6,7 +6,7 @@ test_that("ypr_plot_schedule", {
 })
 
 test_that("ypr_plot_yield", {
-  gp <- ypr_plot_yield(ypr_population())
+  gp <- ypr_plot_yield(ypr_population(), y = "Weight",  pi = seq(0, 1, length.out = 10))
   expect_is(gp, "gg")
 })