+++ NEWS.md
@@ -1,5 +1,6 @@
 - Set default value of `Wa` to be 0.01.
-- Added `Age`, `Length` and `Weight` to `ypr_tabulate_yield()`.
+- Added `Age`, `Length` and `Weight` to `ypr_tabulate_yield()` and added `ypr_tabulate_yields()`
+- Added `y` arg to `ypr_plot_yield()` to allow plotting of mean age, length or weight by capture probability.
 - Added `print.ypr_population()`.
 - Added `ypr_tabulate_parameters()` to get table of parameters, values and descriptions and `ypr_import_parameters()` to convert back to a population object.
 - Added `ypr_tabulate_sr()` to get table of eggs and recruits at unfished, actual and optimal rates.